UnusualSpends/src/test/java/com/amaap/unusualspends/repository/db/impl/FakeInMemoryDatabaseTest.java
package com.amaap.unusualspends.repository.db.impl;

import com.amaap.unusualspends.domain.model.entity.Customer;
import com.amaap.unusualspends.repository.db.InMemoryDatabase;
import com.amaap.unusualspends.repository.impl.InMemoryCustomerRepository;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class FakeInMemoryDatabaseTest {
    FakeInMemoryDatabase fakeInMemoryDatabase = new FakeInMemoryDatabase();
    @Test
    void shouldBeAbleToAddCustomerIntoTable()
    {
        // arrange
        int id = 1;
        String name = "Ashok Pawar";
        String email = "ashokpawar@gmail.com";
        Customer expected = new Customer(id,name,email);

        // act
        Customer actual = fakeInMemoryDatabase.insertIntoCustomerTable(expected);

        // assert
        assertEquals(expected,actual);
    }
}